requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
SQLAlchemy==2.0.25
requests==2.31.0
Pillow==10.2.0
openai==1.13.3

